fixed edit, create a new task and some bugs
