+++ data.js
@@ -1,0 +1,78 @@
+export const arr = [{
+  "iconTask": "idea.png",
+  "name": "test1",
+  "created": "April 8, 2022",
+  "category": "Idea",
+  "content": "text1",
+  "dates": "",
+  "isArchived": false
+},
+{
+  "iconTask": "lateral.png",
+  "name": "test2",
+  "created": "June 13, 2021",
+  "category": "Random Thought",
+  "content": "text2",
+  "dates": "",
+  "isArchived": true
+},
+{
+  "iconTask": "idea.png",
+  "name": "test3",
+  "created": "May 18, 2020",
+  "category": "Idea",
+  "content": "text3",
+  "dates": "",
+  "isArchived": false
+},
+{
+  "iconTask": "shopping-cart.png",
+  "name": "test4",
+  "created": "July 22, 2021",
+  "category": "Task",
+  "content": "text4",
+  "dates": "",
+  "isArchived": true
+},
+{
+  "iconTask": "shopping-cart.png",
+  "name": "test5",
+  "created": "April 8, 2022",
+  "category": "Task",
+  "content": "text5",
+  "dates": "",
+  "isArchived": false
+},
+{
+  "iconTask": "lateral.png",
+  "name": "test6",
+  "created": "May 18, 2020",
+  "category": "Random Thought",
+  "content": "text6",
+  "dates": "",
+  "isArchived": true
+},
+{
+  "iconTask": "idea.png",
+  "name": "test7",
+  "created": "July 22, 2021",
+  "category": "Idea",
+  "content": "text7",
+  "dates": "",
+  "isArchived": true
+}]
+
+export const categoryList = ["Task", "Random Thought", "Idea"]
+
+export const regex = /(3[01]|[12][0-9]|0?[1-9])[\/.-](1[0-2]|0?[1-9])[\/.-](?:[0-9]{2})?[0-9]{2}/gm
+
+
+export const obj = {
+  "inputState": false,
+  "who": "",
+  "nameValue": "",
+  "categoryValue": "",
+  "contentValue": "",
+  "indexItem": ""
+}
+
