+++ index.html
@@ -38,6 +38,7 @@
     <div>
       <input type="button" class="button-create-task" id="button-create-task" value="Create Task"></input>
     </div>
+
     <div id="info-header" class="info-header invert">
       <div class="note-category">
         <span>Note Category</span>
@@ -56,7 +57,7 @@
     <div id="popup" style="display:none">
       <div class="popup-content">
         <img id="cancel" src="./icon/cancel.png" alt="X">
-        
+
         <div class="header invert">
           <div class="name-header">
             <span>Name</span>
@@ -73,7 +74,8 @@
           <div class="dates">
             <span>Dates</span>
           </div>
-         
+          <div class="icon-header-blok">
+          </div>
         </div>
 
         <div id="tasks-popup">
@@ -81,7 +83,6 @@
 
       </div>
     </div>
-
 
   </div>
 </body>
